requirements.txt
Django==4.2.5
django_environ==0.11.2
djangorestframework==3.14.0
djangorestframework_api_key==3.0.0
djangorestframework_simplejwt==5.3.1
Requests==2.32.3
